+++ src/main/java/be/ugent/idlab/knows/dataio/compression/Compression.java
@@ -1,0 +1,11 @@
+package be.ugent.idlab.knows.dataio.compression;
+
+public enum Compression {
+    None,
+    GZip,
+    Zip,
+    XZ,
+    Tar,
+    TarXZ,
+    TarGZ
+}