+++ src/main/java/be/ugent/idlab/knows/dataio/record/CSVRecord.java
@@ -1,6 +1,11 @@
 package be.ugent.idlab.knows.dataio.record;
 
+import com.opencsv.CSVWriter;
+
+import java.io.IOException;
+import java.io.StringWriter;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -107,5 +112,30 @@ public class CSVRecord extends Record {
         int result = data.hashCode();
         result = 31 * result + datatypes.hashCode();
         return result;
+    }
+
+    /**
+     * Writes this CSV record as a string, including headers.
+     * @return This CSV record as a CSV-serialized String.
+     * @throws IOException When writing to CSV goes wrong.
+     */
+    public String toCSVString() throws IOException {
+        int size = this.data.size();
+        String[] header = new String[size];
+        String[] data = new String[size];
+
+        List<String> keys = this.data.keySet().stream().toList();
+
+        for (int i = 0; i < size; i++) {
+            String key = keys.get(i);
+            header[i] = key;
+            data[i] = this.data.get(key);
+        }
+
+        StringWriter sw = new StringWriter();
+        try (CSVWriter writer = new CSVWriter(sw)){
+            writer.writeAll(List.of(header, data));
+        }
+        return sw.toString();
     }
 }