+++ src/main/java/be/ugent/idlab/knows/dataio/record/JSONRecord.java
@@ -1,9 +1,10 @@
 package be.ugent.idlab.knows.dataio.record;
 
 import com.fasterxml.jackson.databind.node.ValueNode;
-import com.jayway.jsonpath.JsonPath;
-import com.jayway.jsonpath.JsonPathException;
-import com.jayway.jsonpath.PathNotFoundException;
+import com.jayway.jsonpath.*;
+import com.jayway.jsonpath.spi.json.JsonProvider;
+import com.jayway.jsonpath.spi.mapper.MappingProvider;
+import net.minidev.json.JSONArray;
 import org.jsfr.json.compiler.JsonPathCompiler;
 import org.jsfr.json.path.PathOperator;
 import org.slf4j.Logger;
@@ -12,6 +13,7 @@ import org.slf4j.LoggerFactory;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -20,22 +22,72 @@ import java.util.regex.Pattern;
  * Every source corresponds with a JSON object in a data source.
  */
 public class JSONRecord extends Record {
-    private final Object document; // JSON object
-    private final String tag; // what iterator was used to obtain this.document
-    private final String path; // what specific path was taken to arrive at this.document
+    // Configuration for Jayway to provide stable outputs
+    static {
+        Configuration.setDefaults(new Configuration.Defaults() {
 
+            private final JsonProvider jsonProvider = Configuration.defaultConfiguration().jsonProvider();
+            private final MappingProvider mappingProvider = Configuration.defaultConfiguration().mappingProvider();
+
+            @Override
+            public JsonProvider jsonProvider() {
+                return jsonProvider;
+            }
+
+            @Override
+            public Set<Option> options() {
+                return Set.of(
+                        Option.DEFAULT_PATH_LEAF_TO_NULL,
+                        Option.REQUIRE_PROPERTIES,
+                        Option.ALWAYS_RETURN_LIST
+                );
+            }
+
+            @Override
+            public MappingProvider mappingProvider() {
+                return mappingProvider;
+            }
+        });
+    }
+
+    private final Object document; // JSON object
+    private final String tag; // the iterator used to obtain this.document
+    private final String path; // the specific path taken to arrive at this.document
+    private final int index; // the index in an array this element is at
     protected Logger logger = LoggerFactory.getLogger(this.getClass());
     private List<String> compiledPath;
 
-
+    /**
+     * Creates a JSONRecord, containing a (sub)document of the original.
+     * @param document  The JSON object obtained.
+     * @param tag       The high level JSONPath (`iterator` in RML speak) used to obtain this document object.
+     */
     public JSONRecord(Object document, String tag) {
         this(document, tag, "");
     }
 
+    /**
+     * Creates a JSONRecord, containing a (sub)document of the original.
+     * @param document  The JSON object obtained.
+     * @param tag       The high level JSONPath (`iterator` in RML speak) used to obtain this document object.
+     * @param path      A sub-path used to obtain this document object. E.g. the last segment of a JSONPath
+     */
     public JSONRecord(Object document, String tag, String path) {
+        this(document, tag, path, -1);
+    }
+
+    /**
+     * Creates a JSONRecord, containing a (sub)document of the original.
+     * @param document  The JSON object obtained.
+     * @param tag       The high level JSONPath (`iterator` in RML speak) used to obtain this document object.
+     * @param path      A sub-path used to obtain this document object. E.g. the last segment of a JSONPath
+     * @param index     If this record is an element of an array, the index of the record in that array.
+     */
+    public JSONRecord(Object document, String tag, String path, int index) {
         this.document = document;
         this.tag = tag;
         this.path = path;
+        this.index = index;
     }
 
     /**
@@ -74,8 +126,13 @@ public class JSONRecord extends Record {
         }
 
         try {
-            Object result = JsonPath.read(this.document, reference);
-            if (result != null) {
+            JSONArray result = JsonPath.read(this.document, reference);
+            boolean allNulls = true;
+            for (Object o : result) {
+                allNulls = allNulls && o == null;
+            }
+
+            if (!result.isEmpty() && !allNulls) {
                 return RecordValue.ok(result);
             } else {
                 return RecordValue.empty();
@@ -178,6 +235,10 @@ public class JSONRecord extends Record {
 
     public String getTag() {
         return tag;
+    }
+
+    public int getIndex() {
+        return index;
     }
 
     @Override