+++ src/test/java/be/ugent/idlab/knows/dataio/access/LocalAccessTest.java
@@ -1,5 +1,6 @@
 package be.ugent.idlab.knows.dataio.access;
 
+import be.ugent.idlab.knows.dataio.compression.Compression;
 import be.ugent.idlab.knows.dataio.cores.LocalAccessTestCore;
 import org.apache.commons.lang3.SystemUtils;
 import org.junit.Assume;
@@ -7,9 +8,12 @@ import org.junit.Test;
 import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.junit.jupiter.api.condition.OS;
 
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.TransformerException;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
@@ -21,7 +25,7 @@ import static org.junit.Assert.assertThrows;
 public class LocalAccessTest extends LocalAccessTestCore {
 
     @Test
-    public void relativePath() throws IOException, SQLException {
+    public void relativePath() throws IOException {
         String relative = "src/test/resources/local_access/file.txt";
 
         Access access = new LocalFileAccess("", relative, "txt", StandardCharsets.UTF_8);
@@ -32,7 +36,7 @@ public class LocalAccessTest extends LocalAccessTestCore {
     }
 
     @Test
-    public void absolutePath() throws IOException, SQLException {
+    public void absolutePath() throws IOException {
         String relative = "src/test/resources/local_access/file.txt";
         String absolute = new File(relative).getAbsolutePath();
         Access access = new LocalFileAccess(absolute, "", "txt", StandardCharsets.UTF_8);
@@ -44,7 +48,7 @@ public class LocalAccessTest extends LocalAccessTestCore {
     }
 
     @Test
-    public void relativeToBase() throws IOException, SQLException {
+    public void relativeToBase() throws IOException {
         String base = "src/test/resources/";
         String baseAbsolute = new File(base).getAbsolutePath();
         Access access = new LocalFileAccess("local_access/file.txt", baseAbsolute, "txt", StandardCharsets.UTF_8);
@@ -86,5 +90,151 @@ public class LocalAccessTest extends LocalAccessTestCore {
         } catch (FileNotFoundException e) {
             throw new FileNotFoundException(e.getMessage());
         }
+    }
+
+    @Test
+    public void compressionGZip() throws SQLException, IOException, ParserConfigurationException, TransformerException {
+        Access access = new LocalFileAccess("access/compression/Friends.json.gz", "src/test/resources", "gzip", Charset.defaultCharset(), Compression.GZip);
+        String expected = """
+                [
+                  {\s
+                    "id": 0,
+                    "name": "Monica Geller",
+                    "age": 33
+                  },
+                  {\s
+                    "id": 1,
+                    "name": "Rachel Green",
+                    "age": 34
+                  },
+                  {\s
+                    "id": 2,
+                    "name": "Joey Tribbiani",
+                    "age": 35
+                  },
+                  {\s
+                    "id": 3,
+                    "name": "Chandler Bing",
+                    "age": 36
+                  },
+                  {\s
+                    "id": 4,
+                    "name": "Ross Geller",
+                    "age": 37
+                  }
+                ]
+                """;
+
+        String actual = new String(access.getInputStream().readAllBytes());
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void compressionZIP() throws SQLException, IOException, ParserConfigurationException, TransformerException {
+        Access access = new LocalFileAccess("access/compression/Friends.json.zip", "src/test/resources", "zip", Charset.defaultCharset(), Compression.Zip);
+        String expected = """
+                [
+                  {\s
+                    "id": 0,
+                    "name": "Monica Geller",
+                    "age": 33
+                  },
+                  {\s
+                    "id": 1,
+                    "name": "Rachel Green",
+                    "age": 34
+                  },
+                  {\s
+                    "id": 2,
+                    "name": "Joey Tribbiani",
+                    "age": 35
+                  },
+                  {\s
+                    "id": 3,
+                    "name": "Chandler Bing",
+                    "age": 36
+                  },
+                  {\s
+                    "id": 4,
+                    "name": "Ross Geller",
+                    "age": 37
+                  }
+                ]
+                """;
+        String actual = new String(access.getInputStream().readAllBytes());
+        assertEquals(expected, actual);
+
+    }
+
+    @Test
+    public void compressionTarGZ() throws SQLException, IOException, ParserConfigurationException, TransformerException {
+        Access access = new LocalFileAccess("access/compression/Friends.json.tar.gz", "src/test/resources", "tar.gz", Charset.defaultCharset(), Compression.TarGZ);
+        String expected = """
+                [
+                  {\s
+                    "id": 0,
+                    "name": "Monica Geller",
+                    "age": 33
+                  },
+                  {\s
+                    "id": 1,
+                    "name": "Rachel Green",
+                    "age": 34
+                  },
+                  {\s
+                    "id": 2,
+                    "name": "Joey Tribbiani",
+                    "age": 35
+                  },
+                  {\s
+                    "id": 3,
+                    "name": "Chandler Bing",
+                    "age": 36
+                  },
+                  {\s
+                    "id": 4,
+                    "name": "Ross Geller",
+                    "age": 37
+                  }
+                ]
+                """;
+        String actual = new String(access.getInputStream().readAllBytes());
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void compressionTarXZ() throws SQLException, IOException, ParserConfigurationException, TransformerException {
+        Access access = new LocalFileAccess("access/compression/Friends.json.tar.xz", "src/test/resources", "tar.xz", Charset.defaultCharset(), Compression.TarXZ);
+        String expected = """
+                [
+                  {\s
+                    "id": 0,
+                    "name": "Monica Geller",
+                    "age": 33
+                  },
+                  {\s
+                    "id": 1,
+                    "name": "Rachel Green",
+                    "age": 34
+                  },
+                  {\s
+                    "id": 2,
+                    "name": "Joey Tribbiani",
+                    "age": 35
+                  },
+                  {\s
+                    "id": 3,
+                    "name": "Chandler Bing",
+                    "age": 36
+                  },
+                  {\s
+                    "id": 4,
+                    "name": "Ross Geller",
+                    "age": 37
+                  }
+                ]
+                """;
+        String actual = new String(access.getInputStream().readAllBytes());
+        assertEquals(expected, actual);
     }
 }